Dev#2

+++ src/app/resourceMgt/resource-edit-form/resource-edit-form.component.ts
@@ -12,6 +12,7 @@ import { OrgUnitService } from '../../shared/sevices_resourceMgt/orgUnit.service
 import { ActivatedRoute, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { NgxSpinnerService } from 'ngx-spinner';
+import { ConfirmDialogService } from '../../ConfirmDialogBox/confirm-dialog.service';
 
 @Component({
   selector: 'app-resource-edit-form',
@@ -26,16 +27,16 @@ export class ResourceEditFormComponent implements OnInit{
   formValue!: FormGroup; 
   selectedResource: any; 
   resourceObject: any;
-  router: any;
 
   constructor(private formBuilder: FormBuilder, 
               private http: HttpClient, 
               private resourceService: ResourceService, 
               private jobRoleService: JobRoleService, 
               private orgUnitService: OrgUnitService,
-              private route: ActivatedRoute,
+              private router: Router,
               private toaster: ToastrService,
-              private spinner: NgxSpinnerService) { }
+              private spinner: NgxSpinnerService,
+              private confirmMessage: ConfirmDialogService) { }
 
   ngOnInit(): void {
     this.loadJobRoles();// calling the loadJobRoles Method
@@ -75,7 +76,7 @@ export class ResourceEditFormComponent implements OnInit{
       })
     )
     .subscribe((res: any) => {
-      debugger;
+      // debugger;
       this.jobroles = res; // Assuming the response is directly the array of resources
       this.spinner.hide();
     },
@@ -96,7 +97,6 @@ export class ResourceEditFormComponent implements OnInit{
       })
     )
     .subscribe((res: any) => {
-      debugger;
       this.orgunits = res; // Assuming the response is directly the array of resources
     },
       (error) => {
@@ -110,12 +110,12 @@ export class ResourceEditFormComponent implements OnInit{
 
   onEditResource(data: ResourceModel) {
     console.log(data);
-    this.resourceService.updateResource(this.selectedResource.resourceId, data)
+    this.confirmMessage.open('Are you sure you want to edit this resource?').subscribe(confirmed => {
+      if (confirmed) {
+        this.resourceService.updateResource(this.selectedResource.resourceId, data)
       .subscribe(
         (res: any) => {
-          debugger;
           console.log('Resource updated successfully:', res);
-          // alert('Resource updated Successfully');
           this.editSucceseMassege(this.selectedResource.resourceId);
           this.formValue.reset();
           this.resourceService.resourceListUpdated.emit(); // Emit the event
@@ -126,10 +126,15 @@ export class ResourceEditFormComponent implements OnInit{
           //display an error message to the user.
         }
       );
+      }
+    });
+    
   }
   
   onDeleteResource() {
-    this.resourceService.deleteResource(this.selectedResource.resourceId)
+  this.confirmMessage.open('Are you sure you want to delete this resource?').subscribe(confirmed => {
+    if(confirmed){
+      this.resourceService.deleteResource(this.selectedResource.resourceId)
     .subscribe((res:ResourceModel)=> {
       console.log('Resource deleted successfully:', res);
       this.deleteSucceseMassege(this.selectedResource.resourceId);
@@ -142,6 +147,12 @@ export class ResourceEditFormComponent implements OnInit{
       // Handle error appropriately, such as displaying an error message to the user.
     }
     );
+    }
+  });
+  }
+
+  onCancel() {
+    this.router.navigate(['pages-body/first-view']);
   }
 
   //Delete Success Message
